Complete the SessionController and implement the SessionExerciseService, implement the GlobalExceptionHandler and fix some bugs and issues in some parts of the code.

+++ ta/src/main/java/com/trainingapp/ta/services/exerciseService/ExerciseServiceImpl.java
@@ -34,7 +34,9 @@ public class ExerciseServiceImpl implements ExerciseService {
 
     @Override
     public void deleteExercise(Long id) {
-        Exercise exercise = exerciseRepository.findById(id).orElseThrow(() -> new ExerciseNotFoundException(id));
+        if (!exerciseRepository.existsById(id)) {
+            throw new ExerciseNotFoundException(id);
+        }
         exerciseRepository.deleteById(id);
     }
